refactor: move validation interface to presentation layer
