+++ src/main/factory/login/login-validation.spec.ts
@@ -1,4 +1,4 @@
-import { IValidation } from '../../../domain/usecases/users/validation'
+import { IValidation } from '../../../presentation/protocols/validation'
 import { ValidationComposite } from '../../../presentation/validations/validation-composite'
 import { RequiredFieldValidation } from '../../../presentation/validations/required-field-validation'
 import { makeLoginValidation } from './login-validation'
