+++ src/main/factory/signup/signup-validation.spec.ts
@@ -1,4 +1,4 @@
-import { IValidation } from '../../../domain/usecases/users/validation'
+import { IValidation } from '../../../presentation/protocols/validation'
 import { ValidationComposite } from '../../../presentation/validations/validation-composite'
 import { RequiredFieldValidation } from '../../../presentation/validations/required-field-validation'
 import { makeSignUpValidation } from './signup-validation'
