+++ src/main/factory/signup/signup-validation.ts
@@ -1,4 +1,4 @@
-import { IValidation } from '../../../domain/usecases/users/validation'
+import { IValidation } from '../../../presentation/protocols/validation'
 import { CompareFieldsValidation } from '../../../presentation/validations/compare-fields-validation'
 import { EmailValidation } from '../../../presentation/validations/email-validation'
 import { RequiredFieldValidation } from '../../../presentation/validations/required-field-validation'
