+++ src/presentation/controller/login/login-controller.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { IEncrypter } from '../../../data/protocols/criptography/encrypter'
 import { IAuthentication } from '../../../domain/usecases/users/authentication'
-import { IValidation } from '../../../domain/usecases/users/validation'
+import { IValidation } from '../../protocols/validation'
 import {
   badRequest,
   Controller,
