+++ src/presentation/validations/email-validation.spec.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { IValidation } from '../../domain/usecases/users/validation'
+import { IValidation } from '../protocols/validation'
 import { InvalidParamError } from '../errors/invalid-param-error'
 import { IEmailValidator } from '../protocols/email-validator'
 import { EmailValidation } from './email-validation'
