+++ src/presentation/validations/validation-composite.spec.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { IValidation } from '../../domain/usecases/users/validation'
+import { IValidation } from '../protocols/validation'
 import { ValidationComposite } from './validation-composite'
 
 interface SutTypes {
